Reimplement pause functionality and refine cursor tracking logic
Now, the cursor is only tracked if the window is in focus and the window is explicitly set to track the mouse. Currently this is updated based on whether the control manager is in the pause state or not. 

The cursor isn't visible when it's over the window however, regardless of whether it's tracked or not.