Implement Alternate Widget. Hide COR when bouncing is disabled
The Alternate widget allows for swapping between two widgets based on either a boolean, or the state of an internal checkbox. It can also be used to hide a widget based on some other value. 

For demonstration I've made the settings hide the CoR slider when bouncing is disabled, instead printing a message telling the user to enable bouncing if they want to edit the CoR.