Implement several random number engines
Previously the random number generator was mapping the base output of the random device to a normal distribution. Now it's capable using the predefined random number generators provided by the standard library. 

These are all statically defined for now, and it's impossible to change the genrator at runtime, but I have plans.