Add framework for new random number system, generator logic
At this point it lacks the ability to change the random number generator engine, however that will change soon. 

The generator class is a part of the new random number generator system. It will provide the same functionality as random manager's RandomNumberEngine class with much less code and more flexibility.